ベクトル更新処理で複数アイテムを更新する際の処理分割を行う

+++ LibPythonAI/python_lib/ai_chat_explorer/ai_app.py
@@ -166,13 +166,15 @@ def update_collection(openai_props: OpenAIProps, vector_db_items: list[VectorDBI
 
 def delete_embeddings(openai_props: OpenAIProps ,vector_db_props: VectorDBItem):
     vector_db = LangChainVectorDB.get_vector_db(openai_props, vector_db_props)
-    for entry in vector_db_props.VectorMetadataList:
+    entry = vector_db_props.VectorMetadata
+    if entry is not None:
         vector_db.delete_document(entry.source_id)
 
 def update_embeddings(openai_props: OpenAIProps ,vector_db_props: VectorDBItem):
     # LangChainVectorDBを生成
     vector_db = LangChainVectorDB.get_vector_db(openai_props, vector_db_props)
-    for entry in vector_db_props.VectorMetadataList:
+    entry = vector_db_props.VectorMetadata
+    if entry is not None:
         vector_db.update_document(entry)
 
 # export_to_excelを実行する
